+++ bocas/artifacts/artifact.py
@@ -1,4 +1,4 @@
-from bocas.artifacts.yaml_utils import parse_yaml_node
+from bocas.yaml_utils import parse_yaml_node
 
 
 class Artifact:
